+++ packages/bakit/CHANGELOG.md
@@ -1,5 +1,29 @@
 # bakit
 
+## 2.1.1
+
+### Patch Changes
+
+- @bakit/service:
+  - Added lifecycle hooks for `initialize` and `ready`.
+  - Added `ready` state checking for transport and driver.
+
+  @bakit/rest:
+  - Removed REST transport proxy.
+  - @bakit/service is no longer used.
+
+  @bakit/gateway:
+  - @bakit/service is no longer used.
+
+  bakit:
+  - TSX is no longer required.
+  - Added a check for using TSX if possible.
+
+- Updated dependencies
+  - @bakit/service@3.2.0
+  - @bakit/gateway@2.1.4
+  - @bakit/rest@2.1.0
+
 ## 2.1.0
 
 ### Minor Changes